Bytecode representation for literal values.
Added bytecode representation for literal values as well as added
infrustructure to handle disassembly.
Chunks now encode the line number from the original Lox source that
produced a given bytecode instruction (or operand etc.) in that chunk,
with the index of each entry in a chunk's `lines` array corresponding to
the index of the bytecode generated from that line of source Lox.
Minor modifications to .clang-format