Fixed an out of bounds error.
Usage of `size_t` was causing a loop condition to always be met as the loop iterator could
never have the sentinal value -1 and thus the loop could never correctly exit. Usage
`size_t` of also caused index iterator to wrap to max value of `size_t` when initialized
with -1, which occured when there were zero local variables.