Added tracing of `Obj' references and `sweep' phase of GC
`Obj' references are now marked using colours; white when unreached, grey when reached but
children references have not been reached and black when all children
have been reached. This makes it easy to traverse the web of objects
without missing any references nor forgetting where the GC is up to.